requirements.txt
flask
flask-smorest
python-dotenv
marshmallow
flask_sqlalchemy
flask-jwt-extended
passlib
flask-migrate
gunicorn
psycopg2
requests
rq